requirements.txt
streamlit
supabase
python-dotenv
pytz
pandas
numpy
scikit-learn
joblib
apscheduler
plotly
requestsq